pyproject.toml
[project]
name = "todoapi-pt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "httpx>=0.28.1",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.11",
]
